Start a conversation: have a way to run a mux instead
The primary driver for this, at least for me, is that sometimes
I want a mux to be what I serve, but still get `Serve` functionality.

I guess in an ideal world, maybe `kami.Serve()` can take an argument,
and if you provide `nil` then it will use `Handler()`.

ServeMux is the best name I can come up with, while trying to stay
backwards compatible.