Support AppEngine Flexible environment with Go 1.7 changes
With the Go 1.7 changes, the request can be modified in the handler for every request (using `r = r.WithContext(ctx)`) so that if we had middleware to grab the AppEngine context, the new request would cause a panic. I'm not sure if a fix for this will be included with the AppEngine package once it officially supports 1.7, but for those of us currently developing with 1.7 on the Flexible Environment, this is an issue. These build tag changes fixes the issue and alleviates the need for adding middleware for the flexible environment to utilize the proper context throughout the application.